fix model name + update enterprise_id

See merge request liempn13/Do_An_Tot_Nghiep!15

+++ lib/model/position/position_model.dart
@@ -4,7 +4,7 @@ class Positions {
   Positions({this.positionId, this.positionName, this.enterpriseId});
   String? positionId;
   String? positionName;
-  Int8? enterpriseId;
+  int? enterpriseId;
 
   factory Positions.fromJson(Map<String, dynamic> json) => Positions(
       positionId: json["position_id"],
